% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(x, y, censor, model)
}
\arguments{
\item{x}{x}

\item{y}{x}

\item{censor}{censoring indicator}

\item{model}{model type}
}
\description{
Check input arguments to determine which model to use
}
\keyword{internal}
