% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.r
\name{check_param}
\alias{check_param}
\title{check_param}
\usage{
check_param(
  n,
  p,
  ntrees,
  mtry,
  nmin,
  alpha,
  split.gen,
  split.rule,
  nsplit,
  replacement,
  resample.prob,
  importance,
  reinforcement,
  track.obs
)
}
\description{
Check parameters
}
\keyword{internal}
