% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaForest.r
\name{ClaForest}
\alias{ClaForest}
\title{ClaForest}
\usage{
ClaForest(x, y, ncat, obs.w, var.prob, resample.preset, param, ...)
}
\description{
Internal function for fitting classification forest
}
\keyword{internal}
