% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOneTree.r
\name{getOneTree}
\alias{getOneTree}
\title{Print a single tree}
\usage{
getOneTree(x, tree = 1, ...)
}
\arguments{
\item{x}{A fitted RLT object}

\item{tree}{the tree number}

\item{...}{...}
}
\description{
Print a single tree
}
