% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuteRate.r
\name{MuteRate}
\alias{MuteRate}
\title{Muting rate}
\usage{
MuteRate(N, P, speed = NULL, info = FALSE)
}
\arguments{
\item{N}{sample size}

\item{P}{dimension}

\item{speed}{Muting speed: moderate or aggressive}

\item{info}{Whether to output detailed information}
}
\value{
A suggested muting rate
}
\description{
Get the muting rate based on sample size \code{N} and dimension \code{P}. This is an experimental feature. When P is too small, this is not recommended.
}
\examples{
MuteRate(500, 100, speed = "aggressive")
}
