% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.r
\name{my_sample}
\alias{my_sample}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{            samples
}\if{html}{\out{</div>}}}
\usage{
my_sample(Num, min, max)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      testing function
}\if{html}{\out{</div>}}
}
