% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.r
\name{print.importance.RLT}
\alias{print.importance.RLT}
\title{Print Importance Summary}
\usage{
\method{print}{importance.RLT}(x, digits = 4, ...)
}
\arguments{
\item{x}{An \code{importance.RLT} object.}

\item{digits}{Number of digits for formatting. Default: 4.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for \code{importance.RLT} objects.
}
