% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.RLT}
\alias{print.RLT}
\title{Print a RLT object}
\usage{
\method{print}{RLT}(x, ...)
}
\arguments{
\item{x}{A fitted RLT object}

\item{...}{...}
}
\description{
Print a RLT object
}
\examples{
\donttest{
set.seed(42)
x <- matrix(rnorm(100 * 5), ncol = 5)
y <- rowSums(x[, 1:2]) + rnorm(100)
fit <- RLT(x, y, ntrees = 50)
print(fit)
}
}
