% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg_Var_Forest.r
\name{Reg_Var_Forest}
\alias{Reg_Var_Forest}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{            Regression random forest with variance estimation
}\if{html}{\out{</div>}}}
\usage{
Reg_Var_Forest(
  x,
  y,
  testx,
  ntrees = 10000,
  mtry = max(1, as.integer(ncol(x)/3)),
  nmin = max(1, as.integer(log(nrow(x)))),
  split.gen = "best",
  nsplit = 0,
  resample.prob = 0.5,
  param.control = list(),
  ncores = 1,
  verbose = 0,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{matrix} or \code{data.frame} of features}

\item{y}{Response variable. a \code{numeric}/\code{factor} vector.}

\item{testx}{A \code{matrix} or \code{data.frame} of testing data}

\item{ntrees}{Number of trees. To obtain a stable and accurate
estimation, the default is \code{ntrees = 10000}.}

\item{mtry}{Number of randomly selected variables used at each
internal node.}

\item{nmin}{Terminal node size. Same as used in \code{RLT}.}

\item{split.gen}{\code{"random"}, \code{"rank"} or \code{"best"}.}

\item{nsplit}{Number of random cutting points to compare for each
variable at an internal node.}

\item{resample.prob}{Proportion of in-bag samples.}

\item{param.control}{A list of additional parameters. However, choices are
limited. For example, reinforcement splitting rules
are not implemented.}

\item{ncores}{Number of cores. Default is 0 (using all available cores).}

\item{verbose}{Whether fitting info should be printed.}

\item{seed}{Random seed number to replicate a previously fitted forest.
Internally, the \verb{xoshiro256++} generator is used. If not specified,
a seed will be generated automatically.}

\item{...}{Additional arguments.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           Prediction and variance estimation of the testing data
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      This function provides some experimental features 
                   to estimate the variance of random forests, especially
                   when the sub-sampling size is larger than \eqn{n/2}. 
                   These trees will always be sampled without replacement.
                   The choices of tuning parameters will be limited.
                   Use at your own risk.
}\if{html}{\out{</div>}}
}
