% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegForest.r
\name{RegForest}
\alias{RegForest}
\title{RegForest}
\usage{
RegForest(x, y, ncat, obs.w, var.w, resample.preset, param, ...)
}
\description{
Internal function for fitting regression forest
}
\keyword{internal}
