% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{set_embed_param}
\alias{set_embed_param}
\title{set_embed_param}
\usage{
set_embed_param(control, reinforcement, n, p)
}
\description{
This is an internal function to set parameters for embedded model.
}
\keyword{internal}
