% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm_gdp.r
\name{skm_gdp_val0_r}
\alias{skm_gdp_val0_r}
\title{skm_gdp_val0_r: selective skm with greedy propgation on val0 cells.
skm_gdp_val0_r: with an input m x n matrix - matrix is init as all rows indentical
and sum to 1 followed by mask a selective columns for each row as 0 - m are stores
sources, n are zipcode desinations and v the cell values are percentage population
lives in the destination, and cells are masked as 0 when s - t within a given dist
X miles, want to select m sequentially so that always covers as much as popluation
objective is sequential max(sum( min(dist(s, t) - all s) < Xmile ? p : 0 - all t))}
\usage{
skm_gdp_val0_r(x)
}
\arguments{
\item{x:}{m x n matrix: s - t - p}
}
\value{
s: index 1 - m gives highest overall p:
  max(sum( min(dist(s, t) - all s) < Xmile ? p : 0 - all t))
}
\description{
skm_gdp_val0_r: selective skm with greedy propgation on val0 cells.
skm_gdp_val0_r: with an input m x n matrix - matrix is init as all rows indentical
and sum to 1 followed by mask a selective columns for each row as 0 - m are stores
sources, n are zipcode desinations and v the cell values are percentage population
lives in the destination, and cells are masked as 0 when s - t within a given dist
X miles, want to select m sequentially so that always covers as much as popluation
objective is sequential max(sum( min(dist(s, t) - all s) < Xmile ? p : 0 - all t))
}

