% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.r
\name{check_input}
\alias{check_input}
\title{check_input}
\usage{
check_input(x, y, censor, model)
}
\arguments{
\item{x}{x}

\item{y}{x}

\item{censor}{censor}

\item{model}{model}
}
\description{
Check input arguments
}
\keyword{internal}
