% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cindex.r
\name{cindex}
\alias{cindex}
\title{C-index}
\usage{
cindex(y, censor, pred, ...)
}
\arguments{
\item{y}{survival time}

\item{censor}{The censoring indicator if survival model is used}

\item{pred}{the predicted value for each subject}
}
\value{
c-index
}
\description{
calculate c-index for survival data
}
