% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.kernel.r
\name{forest.kernel}
\alias{forest.kernel}
\title{\preformatted{      kernel.RLT
}}
\usage{
forest.kernel(
  object,
  X1 = NULL,
  X2 = NULL,
  vs.train = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted RLT object.}

\item{X1}{The the first dataset. This calculates an \eqn{n_1 \times n_1} kernel
matrix of \code{X1}.}

\item{X2}{The the second dataset of the relative kernel weights are required.
If \code{X2} is supplied, then calculate an \eqn{n_1 \times n_2}
kernel matrix. If \code{vs.train} is used, then this must be the original
training data.}

\item{vs.train}{To calculate the kernel weights with respect to the training data.
This is slightly different than supplying the training data to \code{X2}
due to re-samplings of the training process. Hence, \code{ObsTrack} must
be available from the fitted object (using \code{resample.track = TRUE}).}

\item{verbose}{Whether fitting should be printed.}

\item{...}{...   Additional arguments.}
}
\description{
\preformatted{Get random forest induced kernel weight matrix of testing samples 
             or between any two sets of data. This is an experimental feature.
             Use at your own risk.
}
}
