% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.kernel.r
\name{forest.kernel}
\alias{forest.kernel}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{      random forest kernel
}\if{html}{\out{</div>}}}
\usage{
forest.kernel(
  object,
  X1 = NULL,
  X2 = NULL,
  vs.train = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted RLT object.}

\item{X1}{The dataset for prediction. This calculates an \eqn{n_1 \times n_1} kernel
matrix of \code{X1}.}

\item{X2}{The dataset for reference/training.
If \code{X2} is supplied, then calculate an \eqn{n_1 \times n_2}
kernel matrix. If \code{vs.train} is used, then this must be the original
training data.}

\item{vs.train}{To calculate the kernel weights with respect to the training data.
This is slightly different than supplying the training data to \code{X2}
due to re-samplings of the training process. To use this feature, you must
specify \code{resample.track = TRUE} in \code{param.control} when fitting
the forest}

\item{verbose}{Whether fitting should be printed.}

\item{...}{...   Additional arguments.}
}
\value{
A kernel matrix that contains kernel weights for each observation in \code{X1} with respect to \code{X1}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Get random forest induced kernel weight matrix of testing samples 
             or between any two sets of data. This is an experimental feature.
             Use at your own risk.
}\if{html}{\out{</div>}}
}
