% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.kernel.r
\name{forest.kernel}
\alias{forest.kernel}
\title{\preformatted{      kernel.RLT
}}
\usage{
forest.kernel(
  object,
  X1,
  X2 = NULL,
  as.train = FALSE,
  ncores = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted RLT object.}

\item{X1}{The the first dataset.}

\item{X2}{The the second dataset.}

\item{as.train}{If \verb{X2`` is the training data and the kernel weights against the training data is desired. To perform this, the fitted object must contain }ObsTrack\verb{(using}track.obs\verb{). Defult is }FALSE`.}

\item{ncores}{Number of cores. Default is 1.}

\item{verbose}{Whether fitting should be printed.}

\item{...}{...   Additional arguments.}

\item{self}{If \code{TRUE}, then calculate the n by n kernel matrix of \code{X1}.
If \code{FALSE} then calculate the kernel matrix between \code{X1}
and \code{X2}. Default is \code{TRUE}.}
}
\description{
\preformatted{Get random forest induced kernel weights between two sets
             of data.
}
}
