% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.one.tree.r
\name{get.one.tree}
\alias{get.one.tree}
\title{Print a single tree}
\usage{
get.one.tree(x, tree = 1, ...)
}
\arguments{
\item{x}{A fitted RLT object}

\item{tree}{the tree number, starting from 1 to \code{ntrees}.}

\item{...}{...}
}
\description{
Print a single fitted tree from a forest object
}
