% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.surv.band.r
\name{get.surv.band}
\alias{get.surv.band}
\title{get.surv.band}
\usage{
get.surv.band(x, i = 0, alpha = 0.05, ...)
}
\arguments{
\item{x}{A RLT prediction object. This must be an object calculated from a forest
with \code{var.ready = TRUE}.}

\item{i}{Observation number in the prediction, default to calculate all}

\item{alpha}{alpha level for interval \eqn{(\alpha/2, 1 - \alpha/2)}}

\item{...}{...}
}
\description{
Calculate the survival function (two-sided) confidence band from
a RLT survival prediction.
}
