% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.RLT}
\alias{predict.RLT}
\title{prediction using RLT}
\usage{
\method{predict}{RLT}(object, testx = NULL, keep.all = FALSE, ncores = 1, verbose = 0, ...)
}
\arguments{
\item{object}{A fitted RLT object}

\item{testx}{the testing samples, must have the same structure as the training samples}

\item{keep.all}{whether to keep the prediction from all trees}

\item{ncores}{number of cores}

\item{...}{...}
}
\description{
Predict the outcome (regression, classification or survival) using a fitted RLT object
}
