% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.RLT}
\alias{predict.RLT}
\title{prediction using RLT}
\usage{
\method{predict}{RLT}(
  object,
  testx = NULL,
  var.est = FALSE,
  keep.all = FALSE,
  ncores = 1,
  verbose = 0,
  ...
)
}
\arguments{
\item{object}{A fitted RLT object}

\item{testx}{The testing samples, must have the same structure as the
training samples}

\item{var.est}{Whether to estimate the variance of each testing data.
The original forest must be fitted with \code{var.ready = TRUE}.
For survival forests, calculates the covariance matrix over all
observed time points and calculates critical value for the confidence
band.}

\item{keep.all}{whether to keep the prediction from all trees. Warning: this can
occupy a large storage space, especially in survival model}

\item{ncores}{number of cores}

\item{...}{...}
}
\value{
A \code{RLT} prediction object, constructed as a list consisting

\item{Prediction}{Prediction}
\item{Variance}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}}

\strong{For Survival Forests}
\item{hazard}{predicted hazard functions}
\item{CumHazard}{predicted cumulative hazard function}
\item{Survival}{predicted survival function}
\item{Allhazard}{if \code{keep.all = TRUE}, the predicted hazard function for each
observation and each tree}
\item{AllCHF}{if \code{keep.all = TRUE}, the predicted cumulative hazard function for each
observation and each tree}
\item{Cov}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. For each test subject, a matrix of size NFail\eqn{\times}NFail
where NFail is the number of observed failure times in the training data}
\item{Var}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. Marginal variance for each subject}
\item{timepoints}{ordered observed failure times from the training data}
\item{MarginalVar}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. Marginal variance for each subject
from the Cov matrix projected to the nearest positive definite
matrix}
\item{MarginalVarSmooth}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. Marginal variance for each subject
from the Cov matrix projected to the nearest positive definite
matrix and then smoothed using Gaussian kernel smoothing}
\item{CVproj}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. Critical values to calculate confidence bands around
cumulative hazard predictions at several confidence levels. Calculated using
\code{MarginalVar}}
\item{CVprojSmooth}{if \code{var.est = TRUE} and the fitted object is
\code{var.ready = TRUE}. Critical values to calculate confidence bands around
cumulative hazard predictions at several confidence levels. Calculated using
\code{MarginalVarSmooth}}
}
\description{
Predict the outcome (regression, classification or survival)
using a fitted RLT object
}
