% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.RLT}
\alias{predict.RLT}
\title{Prediction function for reinforcement learning trees}
\usage{
\method{predict}{RLT}(object, testx, ...)
}
\arguments{
\item{object}{A fitted RLT object}

\item{testx}{Testing data}

\item{...}{...}
}
\value{
The predicted values. For survival model, it returns the fitted survival functions
}
\description{
Predict future subjects with a fitted RLT model
}
\examples{
x = matrix(rnorm(100), ncol = 10)
y = rowMeans(x)
fit = RLT(x, y, ntrees = 5)
predict(fit, x)
}
