% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.RLT}
\alias{print.RLT}
\title{Print a RLT object}
\usage{
\method{print}{RLT}(x, ...)
}
\arguments{
\item{x}{A fitted RLT object}

\item{...}{...}
}
\description{
Print a RLT object
}
\examples{
x = matrix(rnorm(100), ncol = 10)
y = rowMeans(x)
fit = RLT(x, y, ntrees = 5)
fit
}
