% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuanForest.r
\name{QuanForest}
\alias{QuanForest}
\title{QuanForest}
\usage{
QuanForest(x, y, ncat, obs.w, var.w, resample.preset, param, ...)
}
\description{
Internal function for fitting quantile forest
}
\keyword{internal}
