% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaForest.r, R/RegForest.r
\name{RegForest}
\alias{RegForest}
\alias{ClaForest}
\title{ClaForest}
\usage{
ClaForest(x, y, ncat, obs.w, var.w, resample.preset, param, ...)

RegForest(x, y, ncat, obs.w, var.w, resample.preset, param, ...)
}
\description{
Internal function for fitting classification forest

Internal function for fitting regression forest
}
\keyword{internal}
