% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegForest.r
\name{RegForest}
\alias{RegForest}
\title{RegForest}
\usage{
RegForest(
  x,
  y,
  ncat,
  param,
  RLT.control,
  obs.w,
  var.w,
  ncores,
  verbose,
  ObsTrack,
  ...
)
}
\description{
Internal function for fitting regression forest
}
\keyword{internal}
