% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvForest.r
\name{SurvForest}
\alias{SurvForest}
\title{SurvForest}
\usage{
SurvForest(x, y, censor, ncat, obs.w, var.w, resample.preset, param, ...)
}
\description{
Internal function for fitting survival forest
}
\keyword{internal}
